requirements.txt
astroid
Django
isort
lazy-object-proxy
mccabe
Pillow
pylint
pytz
six
typed-ast
wrapt
qrcode
pywebview